:recycle: move SnowflakeList to `private.py`
