+++ discord/utils/__init__.py
@@ -25,12 +25,8 @@ DEALINGS IN THE SOFTWARE.
 
 from __future__ import annotations
 
-import array
 import collections.abc
-import datetime
 import json
-import re
-from bisect import bisect_left
 from inspect import signature as _signature
 from typing import (
     TYPE_CHECKING,
@@ -40,7 +36,6 @@ from typing import (
     Generic,
     Iterable,
     Iterator,
-    Literal,
     Mapping,
     Protocol,
     Sequence,
@@ -322,37 +317,3 @@ def get_slots(cls: type[Any]) -> Iterator[str]:
             yield from mro.__slots__
         except AttributeError:
             continue
-
-
-class SnowflakeList(array.array):
-    """Internal data storage class to efficiently store a list of snowflakes.
-
-    This should have the following characteristics:
-
-    - Low memory usage
-    - O(n) iteration (obviously)
-    - O(n log n) initial creation if data is unsorted
-    - O(log n) search and indexing
-    - O(n) insertion
-    """
-
-    __slots__ = ()
-
-    if TYPE_CHECKING:
-
-        def __init__(self, data: Iterable[int], *, is_sorted: bool = False): ...
-
-    def __new__(cls, data: Iterable[int], *, is_sorted: bool = False):
-        return array.array.__new__(cls, "Q", data if is_sorted else sorted(data))  # type: ignore
-
-    def add(self, element: int) -> None:
-        i = bisect_left(self, element)
-        self.insert(i, element)
-
-    def get(self, element: int) -> int | None:
-        i = bisect_left(self, element)
-        return self[i] if i != len(self) and self[i] == element else None
-
-    def has(self, element: int) -> bool:
-        i = bisect_left(self, element)
-        return i != len(self) and self[i] == element
